Tables clear code, allow to change table count if all tables are available

+++ app/Http/Controllers/TablesController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use App\Http\Requests\CreateTablesRequest;
 use App\Models\Tables;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 
 class TablesController extends Controller
 {
@@ -17,17 +16,8 @@ class TablesController extends Controller
     public function index()
     {
         $tables = count(Tables::all());
-        return view('tables.index', compact('tables'));
-    }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
+        return view('tables.index', compact('tables'));
     }
 
     /**
@@ -39,60 +29,25 @@ class TablesController extends Controller
     public function store(CreateTablesRequest $request)
     {
         $number_of_tables = $request->number_of_tables;
+
+        if (self::checkForOccupiedTables() !== 0) {
+
+            return redirect()->route('table.index')->with('error', 'Please close all orders. Then you can change table numbers.');
+        }
+
         Tables::truncate();
+
         for ($i = 0; $i < $number_of_tables; $i++) {
             Tables::insert([
                 'is_available' => true,
             ]);
         }
 
-
         return redirect()->route('table.index')->with('success', 'Successful updated number ot tables and zones');
-
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
+    public function checkForOccupiedTables()
     {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request $request
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
-    {
-        //
+        return Tables::select('is_available')->where('is_available', false)->count();
     }
 }
